Update landing page seeder for hotel installation, adjust demo data configuration, and ensure proper type casting for reservation calculations.

+++ app/Modules/Reservation/Actions/CreateReservation.php
@@ -99,7 +99,7 @@ class CreateReservation
 
     protected function estimateTotal(ReservationData $data, CarbonImmutable $checkIn, CarbonImmutable $checkOut): int
     {
-        $nights = max(1, $checkIn->diffInDays($checkOut));
+        $nights = max(1, (int) $checkIn->diffInDays($checkOut));
         $nightly = 0;
 
         if ($data->bedId !== null) {
@@ -116,6 +116,6 @@ class CreateReservation
         $subtotal = $nightly * $nights;
         $afterDiscount = max(0, $subtotal - $data->discount);
 
-        return $afterDiscount + $data->tax;
+        return (int) ($afterDiscount + $data->tax);
     }
 }
